Edited parse data function, need to fix issues. Removed main activity log messages.

+++ app/src/main/java/com/example/sdp11/wmd/DataFragment.java
@@ -35,11 +35,15 @@ public class DataFragment extends Fragment implements View.OnClickListener, Swip
     @Override
     public void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        dataSource = MainActivity.dataSource;
+
 
         //Test
-        dataSource.deleteAllThrows();
-        addDemoThrows();
+        if(dataSource == null) {
+            dataSource = MainActivity.dataSource;
+            dataSource.deleteAllThrows();
+            addDemoThrows();
+        }
+
     }
 
     @Override
